\name{sl2}
\alias{sl2}
\alias{sl2names}
\alias{sl2seq}
\alias{sl2seqnames}
\alias{bg3}
\alias{bg3names}
\docType{data}
\title{
  Sample binding site and related data from Drosophila melanogaster.
}
\description{
  chroGPS example ChIP-CHIP data for S2 and BG3 cell lines (modEncode),
  ChIP-Seq data for S2 cell lines (NCBI GEO GSE19325).
}
\usage{data(sl2)}
\source{
  \code{http://www.modencode.org}
  \code{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19325}
}
\references{
  \code{http://www.modencode.org}
  \code{http://www.ncbi.nlm.nih.gov/geo/}
}
\examples{
data(sl2)
}
\keyword{datasets}
